# Always use implementation 2 in GRU and LSTM for performance