PERF

Also in this change:
- Fix a small bug in the logic of GRU dropout
- Add more unit tests to ensure faithful preservation of `implementation` in the config for GRU and LSTM, even though it is ignored in terms of actual implementation in TensorFlow.js

Benchmark results:
- From integration_tests/benchmark:
    - GRU
      - predict: 106 ms --> 83.2 ms (1.27x)
      - fit: 180 ms --> 106 ms (1.7x)
    - LSTM
      - predict: 137.7 ms --> 64.3 ms (2.1x)
      - fit: 224.0 ms --> 100 ms (2.24x)
    - Attention
      - predict: 148.9 ms --> 77.9 ms (1.9x)

Similar speed-up is observed for tfjs-node:
- LSTM input shape: [null, 100, 128]:  predict:1226.7 ms --> 602.2 ms (2.0x)

Relating to https://github.com/tensorflow/tfjs/issues/863
<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-layers/364)
<!-- Reviewable:end -->